// ui-packages/ui-components/lib/sidenav/sidenav.type.ts
export enum FileMenuItemIds {
  Rename = 'Rename',
  Delete = 'Delete',
}

export enum NewFileMenuItemIds {
  AddSketchFile = 'AddSketchFile',
  AddHeaderFile = 'AddHeaderFile',
  AddTextFile = 'AddTextFile',
  AddSecretsTab = 'AddSecretsTab',
  ImportFile = 'ImportFile',
}

export type FileMenuSection = DropdownMenuSectionType<
  NewFileMenuItemIds | FileMenuItemIds,
  MessageDescriptor
>;
export type FileMenuItemType = DropdownMenuItemType<
  NewFileMenuItemIds | FileMenuItemIds,
  MessageDescriptor
>;

export type FileMenuType = (
  fileId: string,
  fileName: string,
  fileExtension: string,
) => void;

export type FileMenuDictionary<T> = { [K in FileMenuItemIds]: T };
export type FileMenuItemDictionary = FileMenuDictionary<FileMenuItemType>;
export type FileMenuHandlerDictionary = FileMenuDictionary<FileMenuType>;

export type NewFileMenuDictionary<T> = { [K in NewFileMenuItemIds]: T };
export type NewFileMenuItemDictionary = NewFileMenuDictionary<FileMenuItemType>;